complete documentation (1/2)
